Add data.rs
