+++ scrypto/src/component/mod.rs
@@ -9,7 +9,4 @@ pub use component::{
 };
 pub use kv_store::{KeyValueStore, ParseKeyValueStoreError};
 pub use package::{BorrowedPackage, Package, PackageAddress, PackagePublishInput};
-pub use system::{
-    component_system, init_component_system, ComponentSystem, DataValue, DataValueRef,
-    DataValueRefMut,
-};
+pub use system::{component_system, init_component_system, ComponentSystem, DataValue};
