+++ scrypto/src/core/mod.rs
@@ -1,4 +1,5 @@
 mod actor;
+mod data;
 mod invocation;
 mod level;
 mod logger;
@@ -6,6 +7,7 @@ mod network;
 mod runtime;
 
 pub use actor::{ScryptoActor, ScryptoActorInfo};
+pub use data::*;
 pub use invocation::{ComponentOffset, DataAddress, SNodeRef};
 pub use level::Level;
 pub use logger::Logger;
